src/requirements.dev.txt
flake8
pylint
pylint-flask
autopep8
pytest
coverage